Crate sc_consensus_nakamoto

Structs

BitcoinBlockImporter

Bitcoin specific block import implementation.

BlockImportQueue

Subcoin import queue for processing Bitcoin blocks.

BlockVerifier

A struct responsible for verifying Bitcoin blocks.

ChainParams

Extended [Params].

HeaderVerifier

A struct responsible for verifying block header.

ImportBlocks

Represents a batch of Bitcoin blocks that are to be imported.

ImportConfig

Block import configuration.

ImportManyBlocksResult

Result of import_many_blocks.

SubstrateImportQueueVerifier

Verifier used by the Substrate import queue.

Enums

BlockVerification

Represents the level of block verification.

Error

Consensus error type.

HeaderError

Block header error.

ImportStatus

Result of the operation of importing a Bitcoin block.

ScriptEngine

Represents the Bitcoin script backend.

Traits

BitcoinBlockImport

A trait for importing Bitcoin blocks.

Functions

bitcoin_import_queue

Creates a new import queue.