Crate sc_consensus_nakamoto

Structs

BenchmarkAllExecutor

BenchmarkRuntimeBlockExecutor

BitcoinBlockImporter

Bitcoin specific block import implementation.

BlockImportQueue

Import queue for processing Bitcoin blocks.

BlockVerifier

A struct responsible for verifying Bitcoin blocks.

HeaderVerifier

A struct responsible for verifying block header.

ImportBlocks

Represents a batch of Bitcoin blocks that are to be imported.

ImportConfig

Block import configuration.

ImportManyBlocksResult

Result of import_many_blocks.

OffRuntimeBlockExecutor

Block executor using custom apply_extrinsics, for the initial sync process.

RuntimeBlockExecutor

Block executor using the runtime api execute_block.

Enums

BlockExecutionStrategy

Represents the different strategies for executing a block.

BlockVerification

Represents the level of block verification.

ClientContext

Backend type for the execution client.

Error

ExecutionBackend

Represents the state backend storage type for block execution.

ImportStatus

Result of the operation of importing a Bitcoin block.

Traits

BitcoinBlockImport

A trait for importing Bitcoin blocks.

BlockExecutor

Trait for executing and importing the block.

Functions

bitcoin_import_queue

Creates a new import queue.